Refactor : queries 폴더 삭제
